LayoutTests/platform/chromium/inspector/debugger/function-details-expected.txt

Tests that Debugger.getFunctionDetails command returns correct location. Bug 71808

Debugger was enabled.

Running: testGetFirstLineFunctionDetails
firstLineFunction type = function
Function details: 
lineNumber: 5
columnNumber: 36
scriptId is valid: true
name: firstLineFunction
displayName: undefined
inferredName: undefined

Running: testGetNonFirstLineFunctionDetails
nonFirstLineFunction type = function
Function details: 
lineNumber: 10
columnNumber: 32
scriptId is valid: true
name: notFirstLineFunction
displayName: undefined
inferredName: undefined

Running: testGetDetailsOfFunctionWithInferredName
obj.m type = function
Function details: 
lineNumber: 16
columnNumber: 15
scriptId is valid: true
name: undefined
displayName: undefined
inferredName: obj.m

Running: testGetDetailsOfFunctionWithDisplayName
functionWithDisplayName type = function
Function details: 
lineNumber: 19
columnNumber: 32
scriptId is valid: true
name: functionWithDisplayName
displayName: undefined
inferredName: undefined
Debugger was disabled.
